fix: UserService.java 의 delete 메서드 수정

+++ src/main/java/com/sprint/mission/discodeit/service/UserService.java
@@ -89,9 +89,11 @@ public class UserService {
         }
 
         UserStatus userStatus = userStatusRepository.findByUserId(user.getId())
-                .orElseThrow(() -> new NoSuchElementException("UserStatus Not found"));
+                .orElse(null);
 
+        if (userStatus != null) {
         userStatusRepository.deleteById(userStatus.getId());
+        }
         userRepository.deleteById(user.getId());
     }
